+++ Classes/Controller/ContentElements/SliderContentElementController.php
@@ -1,6 +1,8 @@
 <?php
 namespace T3v\T3vStage\Controller\ContentElements;
 
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+
 use T3v\T3vContent\Controller\ContentElements\ContentElementController;
 
 use T3v\T3vStage\Domain\Repository\SlideRepository;
@@ -25,22 +27,42 @@ class SliderContentElementController extends ContentElementController {
    * @return void
    */
   public function indexAction() {
-    $settings   = $this->settings;
-    $uids       = $settings['slides'];
-    $slides     = $this->slideRepository->findByUids($uids);
-    $responsive = $settings['responsive'];
+    $settings = $this->settings;
 
-    $settings['responsive'] = $this->sortResponsiveSettingsByMinWidth($responsive, 'minWidth');
+    $responsiveSettings = $settings['responsive'];
+
+    if (!empty($responsiveSettings)) {
+      $settings['responsive'] = $this->sortResponsiveSettingsByMinWidth($responsiveSettings);
+    }
+
+    $slides = $this->getSlidesFromSettings($settings);
 
     $this->view->assign('settings', $settings);
     $this->view->assign('slides', $slides);
   }
 
   /**
+   * Helper function to get slides from settings.
+   *
+   * @param array $settings The settings
+   * @return array The slides
+   */
+  protected function getSlidesFromSettings($settings) {
+    $slides     = [];
+    $slidesUids = GeneralUtility::intExplode(',', $settings['slides'], true);
+
+    foreach($slidesUids as $uid) {
+      $slides[] = $this->slideRepository->findByUid($uid);
+    }
+
+    return $slides;
+  }
+
+  /**
    * Helper function to sort the responsive settings by minimum width.
    *
    * @param array $settings The responsive settings
-   * @return array The sorted responsive settings by minimum width
+   * @return array The by minimum width sorted responsive settings
    */
   protected function sortResponsiveSettingsByMinWidth($settings) {
     if (!empty($settings)) {